// ui/src/gen/TimelineFacts.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TimelineActiveTurn } from "./TimelineActiveTurn";
import type { TimelineAgentMessageReceipt } from "./TimelineAgentMessageReceipt";
import type { TimelineLatestContext } from "./TimelineLatestContext";
import type { TimelineLatestPlan } from "./TimelineLatestPlan";
import type { TimelineToolStreamKey } from "./TimelineToolStreamKey";

export type TimelineFacts = { project_root: string, latest_event_id: number, active_turns: Array<TimelineActiveTurn>, latest_contexts: Array<TimelineLatestContext>, latest_agent_messages: Array<TimelineAgentMessageReceipt>, latest_plans: Array<TimelineLatestPlan>, settled_tool_streams: Array<TimelineToolStreamKey>, approval_mode_revision: number, latest_turn_start_id: number | null, steered_message_ids: number[], };